+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/AbstractSpotlessMojo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2024 DiffPlug
+ * Copyright 2016-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -91,6 +91,7 @@ public abstract class AbstractSpotlessMojo extends AbstractMojo {
 
 	static final String GOAL_CHECK = "check";
 	static final String GOAL_APPLY = "apply";
+	static final String GOAL_PRE_PUSH_HOOK = "install-git-pre-push-hook";
 
 	@Component
 	private RepositorySystem repositorySystem;